// src/enums/store.enums.ts
export enum NewsActionEnums {
  FETCH_TOP_HEADLINES = "FETCH_TOP_HEADLINES",
  FETCH_SOURCES = "FETCH_SOURCES",
}

export enum NewsMutationEnums {
  SET_TOP_HEADLINES = "SET_TOP_HEADLINES",
  SET_SOURCES = "SET_SOURCES",
  SET_RECENT = "SET_RECENT",
  SET_CURRENT_SOURCE = "SET_CURRENT_SOURCE",
  SET_CURRENT_COUNTRY = "SET_CURRENT_COUNTRY",
  SET_HEADLINE_FILTER = "SET_HEADLINE_FILTER",
  SET_HEADLINE_LOADING = "SET_HEADLINE_LOADING",
  SET_ARTICLE_HISTORY = "SET_ARTICLE_HISTORY",
}

export enum AppMutationEnums {
  SHOW_NOTIFICATION = "SHOW_NOTIFICATION",
}